+++ sklearn/linear_model/perceptron.py
@@ -2,7 +2,7 @@
 # License: BSD 3 clause
 
 from .stochastic_gradient import BaseSGDClassifier
-from ..feature_selection.selector_mixin import _LearntSelectorMixin
+from ..feature_selection.from_model import _LearntSelectorMixin
 
 
 class Perceptron(BaseSGDClassifier, _LearntSelectorMixin):
